Updated clickable elements logic
Removed the 'edit' buttons and turned the elements themselves into buttons.

This is a test to write a longer commit message instead of an one liner.
In the future these buttons will be stylized in a way that seamlessly integrates them into the UI.
The reason buttons were chosen was because they make more lexical sense as in the functionality, the focus order and their 'clickable' functionalities (onclick event).
Also a few minor bugs were fixed, including an infinite loop with an alert message editing in DisplayTaskName. These minor changes are staged in order to avoid the overload of commits in the early stage of the project.