Created controlled input component for time
This new component has an HTML5 input tag at it's core, and uses the state to control the text to a format of 'HH:MM:SS' as to represent an arbitrary amount of time. It's inner logic will keep track of the position of the user's selection caret and make sure that the input is valid, so that it can lift the state back to handle the update of the interval it represents.
This implementation works with the onKeyDown event, which fires at every key press of the user. However this does not work on most mobile integrated keyboards, which has been a well known issue, so it will be a future task to create an alternate behavior with the onChange event, so that it can be used in mobile.