Implemented CRUD operations on projects
Also integrated the dropdown into the list of tasks, with its features to receive the list of existing projects and display the project that the task belongs to, if any, but also to select a different one or even create a new one. For the rest of the operations a new panel was created.