New system to create example data
Instead of hardcoding example Tasks with all of their values, we now have an array of example Tasks consisting of their name, starting and ending hours, their projects and an offset value that refers to the date to assign to it with respect to the moment these are generated.
This way we can always have Tasks defined in specific ranges to showcase specific behavior, like how to treat Tasks that were generated for today, yesterday, last week, etc.
For this we also created a function to parse the example Tasks created.