Tasks are split by milestone
A milestone is defined as a range in time for which the Tasks are categorized in, like Today, Yesterday, Last Week, etc, so that we can have the list of Tasks a bit more organized.
The component was modified so that it doesn't just lists all of the Tasks right away, but now sorts them by date and then applies the categorization rules to see where do they belong.