Implemented the Modal component
Implemented a new component that will act as a Modal window. It has the necessary logic to be opened on the press of a button and closed when pressed on the close button or anywhere outside of the main modal window.
It also has the logic to ask the user for confirmation before firing a callback, which will be useful when implementing warnings in actions that will benefit from having a confirmation step.