Refactored the whole function
Refactored the whole function by taking it to another file, but also changing how the filters are applied. Now instead of having conditionals hardcoded for each of the filters inside of the function, an array of filters is passed and each one of them is evaluated in chronological order, which makes this function decoupled from the implementation of the milestone system and successfully applies the Open-Closed Principle.
Also improved the documentation